renamed curriculumlist to curriculumtable. Resolved circular dependencies
