--- src/components/CurriculumList/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./CurriculumList";
-export * from "./CurriculumList";
