+++ src/components/CurriculumTable/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./CurriculumTable";
+export * from "./CurriculumTable";
